+++ multinet/src/datastructures/attributes/AttributeStore.cpp
@@ -18,7 +18,7 @@ const vector<AttributeSharedPtr>& MainMemoryAttributeStore::attributes() const {
 	return attribute_vector;
 }
 
-AttributeSharedPtr MainMemoryAttributeStore::attribute(int idx) const {
+AttributeSharedPtr MainMemoryAttributeStore::attribute(size_t idx) const {
 	if (attribute_vector.size()>idx)
 		return attribute_vector.at(idx);
 	else return AttributeSharedPtr();
