+++ multinet/src/generation/UniformEvolutionModel.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 namespace mlnet {
 
-UniformEvolutionModel::UniformEvolutionModel(int m0) {
+UniformEvolutionModel::UniformEvolutionModel(size_t m0) {
 	UniformEvolutionModel::m0 = m0;
 }
